Fix import errors for database module
